+++ lib/src/androidTest/java/com/smileidentity/compose/document/OrchestratedDocumentVerificationScreenTest.kt
@@ -3,6 +3,8 @@ package com.smileidentity.compose.document
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import com.smileidentity.compose.components.LocalMetadata
+import com.smileidentity.compose.nav.ResultCallbacks
 import com.smileidentity.models.JobType
 import com.smileidentity.util.randomUserId
 import com.smileidentity.viewmodel.document.DocumentVerificationViewModel
@@ -21,6 +23,9 @@ class OrchestratedDocumentVerificationScreenTest {
         // when
         composeTestRule.setContent {
             OrchestratedDocumentVerificationScreen(
+                content = {},
+                resultCallbacks = ResultCallbacks(),
+                showSkipButton = false,
                 viewModel = DocumentVerificationViewModel(
                     jobType = JobType.DocumentVerification,
                     userId = randomUserId(),
@@ -29,6 +34,7 @@ class OrchestratedDocumentVerificationScreenTest {
                     countryCode = "254",
                     documentType = "NATIONAL_ID",
                     captureBothSides = false,
+                    metadata = LocalMetadata.current,
                 ),
             )
         }
@@ -45,6 +51,9 @@ class OrchestratedDocumentVerificationScreenTest {
         // when
         composeTestRule.setContent {
             OrchestratedDocumentVerificationScreen(
+                content = {},
+                resultCallbacks = ResultCallbacks(),
+                showSkipButton = false,
                 viewModel = DocumentVerificationViewModel(
                     jobType = JobType.DocumentVerification,
                     userId = randomUserId(),
@@ -53,6 +62,7 @@ class OrchestratedDocumentVerificationScreenTest {
                     countryCode = "254",
                     documentType = "NATIONAL_ID",
                     captureBothSides = false,
+                    metadata = LocalMetadata.current,
                 ),
             )
         }