+++ lib/src/androidTest/java/com/smileidentity/compose/selfie/OrchestratedSelfieCaptureScreenTest.kt
@@ -3,6 +3,7 @@ package com.smileidentity.compose.selfie
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import com.smileidentity.compose.nav.ResultCallbacks
 import org.junit.Rule
 import org.junit.Test
 
@@ -16,7 +17,12 @@ class OrchestratedSelfieCaptureScreenTest {
         val instructionsSubstring = "Next, we'll take a quick selfie"
 
         // when
-        composeTestRule.setContent { OrchestratedSelfieCaptureScreen() }
+        composeTestRule.setContent {
+            OrchestratedSelfieCaptureScreen(
+                content = {},
+                resultCallbacks = ResultCallbacks(),
+            )
+        }
 
         // then
         composeTestRule.onNodeWithText(instructionsSubstring, substring = true).assertIsDisplayed()
@@ -31,6 +37,8 @@ class OrchestratedSelfieCaptureScreenTest {
         composeTestRule.setContent {
             OrchestratedSelfieCaptureScreen(
                 showInstructions = false,
+                content = {},
+                resultCallbacks = ResultCallbacks(),
             )
         }